Refactor Proc to ProcGroup
