+++ cli/daemon/run/run.go
@@ -29,7 +29,6 @@ import (
 	"encore.dev/appruntime/exported/config"
 	"encore.dev/appruntime/exported/experiments"
 	"encr.dev/cli/daemon/apps"
-	"encr.dev/cli/daemon/internal/sym"
 	"encr.dev/cli/daemon/run/infra"
 	"encr.dev/cli/daemon/secret"
 	"encr.dev/cli/internal/xos"
@@ -142,15 +141,15 @@ type RunLogger interface {
 	RunStderr(r *Run, line []byte)
 }
 
-// Proc returns the current running process.
+// ProcGroup returns the current running process.
 // It may have already exited.
 // If the proc has not yet started it may return nil.
-func (r *Run) Proc() *Proc {
-	p, _ := r.proc.Load().(*Proc)
+func (r *Run) ProcGroup() *ProcGroup {
+	p, _ := r.proc.Load().(*ProcGroup)
 	return p
 }
 
-func (r *Run) StoreProc(p *Proc) {
+func (r *Run) StoreProc(p *ProcGroup) {
 	r.proc.Store(p)
 }
 
@@ -215,11 +214,11 @@ func (r *Run) start(ln net.Listener, tracker *optracker.OpTracker) (err error) {
 	// Monitor the running proc and Close the app when it exits.
 	go func() {
 		for {
-			p := r.proc.Load().(*Proc)
+			p := r.proc.Load().(*ProcGroup)
 			<-p.Done()
 			// p exited, but it could have been a reload.
 			// Check to make sure p is still the active proc.
-			p2 := r.proc.Load().(*Proc)
+			p2 := r.proc.Load().(*ProcGroup)
 			if p2 == p {
 				// We're done.
 				for _, ln := range r.Mgr.listeners {
@@ -332,7 +331,7 @@ func (r *Run) buildAndStart(ctx context.Context, tracker *optracker.OpTracker) e
 	}
 
 	startOp := tracker.Add("Starting Encore application", start)
-	newProcess, err := r.StartProc(&StartProcParams{
+	newProcess, err := r.StartProcGroup(&StartProcGroupParams{
 		Ctx:            ctx,
 		BuildDir:       build.Dir,
 		BinPath:        build.Exe,
@@ -357,7 +356,7 @@ func (r *Run) buildAndStart(ctx context.Context, tracker *optracker.OpTracker) e
 
 	previousProcess := r.proc.Swap(newProcess)
 	if previousProcess != nil {
-		previousProcess.(*Proc).Close()
+		previousProcess.(*ProcGroup).Close()
 	}
 
 	tracker.Done(startOp, 50*time.Millisecond)
@@ -365,31 +364,7 @@ func (r *Run) buildAndStart(ctx context.Context, tracker *optracker.OpTracker) e
 	return nil
 }
 
-// Proc represents a running Encore process.
-type Proc struct {
-	ID          string           // unique process id
-	Run         *Run             // the run the process belongs to
-	Pid         int              // the OS process id
-	Meta        *meta.Data       // app metadata snapshot
-	Started     time.Time        // when the process started
-	Experiments *experiments.Set // enabled experiments
-
-	ctx      context.Context
-	log      zerolog.Logger
-	exit     chan struct{} // closed when the process has exited
-	cmd      *exec.Cmd
-	reqWr    *os.File
-	respRd   *os.File
-	buildDir string
-	Client   *yamux.Session
-	authKey  config.EncoreAuthKey
-
-	sym       *sym.Table
-	symErr    error
-	symParsed chan struct{} // closed when sym and symErr are set
-}
-
-type StartProcParams struct {
+type StartProcGroupParams struct {
 	Ctx            context.Context
 	BuildDir       string
 	BinPath        string
@@ -404,11 +379,11 @@ type StartProcParams struct {
 	Experiments    *experiments.Set
 }
 
-// StartProc starts a single actual OS process for app.
-func (r *Run) StartProc(params *StartProcParams) (p *Proc, err error) {
+// StartProcGroup starts a single actual OS process for app.
+func (r *Run) StartProcGroup(params *StartProcGroupParams) (p *ProcGroup, err error) {
 	pid := GenID()
 	authKey := genAuthKey()
-	p = &Proc{
+	p = &ProcGroup{
 		ID:          pid,
 		Run:         r,
 		Experiments: params.Experiments,
@@ -518,72 +493,6 @@ func (r *Run) StartProc(params *StartProcParams) (p *Proc, err error) {
 
 	go p.waitForExit()
 	return p, nil
-}
-
-// Done returns a channel that is closed when the process has exited.
-func (p *Proc) Done() <-chan struct{} {
-	return p.exit
-}
-
-// Close closes the process and waits for it to shutdown.
-// It can safely be called multiple times.
-func (p *Proc) Close() {
-	p.reqWr.Close()
-	timer := time.NewTimer(10 * time.Second)
-	defer timer.Stop()
-	select {
-	case <-p.exit:
-	case <-timer.C:
-		// The process didn't exit after 10s
-		p.log.Error().Msg("timed out waiting for process to exit; killing")
-		p.cmd.Process.Kill()
-		<-p.exit
-	}
-}
-
-func (p *Proc) waitForExit() {
-	defer close(p.exit)
-	defer closeAll(p.reqWr, p.respRd)
-
-	if err := p.cmd.Wait(); err != nil && p.ctx.Err() == nil {
-		p.log.Error().Err(err).Msg("process exited with error")
-	} else {
-		p.log.Info().Msg("process exited successfully")
-	}
-
-	// Flush the logs in case the output did not end in a newline.
-	for _, w := range [...]io.Writer{p.cmd.Stdout, p.cmd.Stderr} {
-		if w != nil {
-			w.(*logWriter).Flush()
-		}
-	}
-}
-
-// parseSymTable parses the symbol table of the binary at binPath
-// and stores the result in p.sym and p.symErr.
-func (p *Proc) parseSymTable(binPath string) {
-	parse := func() (*sym.Table, error) {
-		f, err := os.Open(binPath)
-		if err != nil {
-			return nil, err
-		}
-		defer f.Close()
-		return sym.Load(f)
-	}
-
-	defer close(p.symParsed)
-	p.sym, p.symErr = parse()
-}
-
-// SymTable waits for the proc's symbol table to be parsed and then returns it.
-// ctx is used to cancel the wait.
-func (p *Proc) SymTable(ctx context.Context) (*sym.Table, error) {
-	select {
-	case <-ctx.Done():
-		return nil, ctx.Err()
-	case <-p.symParsed:
-		return p.sym, p.symErr
-	}
 }
 
 // closeAll closes all the given closers, skipping ones that are nil.
